Move button group to stable
