+++ src/stable/components/ButtonGroup/ButtonGroup.js
@@ -1,6 +1,6 @@
 import styles from '!!raw-loader!./ButtonGroup.css';
-import varStyles from '!!raw-loader!../../../../shared/variables.css';
-import bootstrapStyles from '!!raw-loader!../../../../shared/themed-bootstrap.css';
+import varStyles from '!!raw-loader!../../../shared/variables.css';
+import bootstrapStyles from '!!raw-loader!../../../shared/themed-bootstrap.css';
 
 const template = document.createElement('template');
