+++ src/stable/index-stable.js
@@ -3,3 +3,4 @@ import './components/SectionNavigation/cod-section-navigation.js';
 import './components/ServiceButton/cod-service-button.js';
 import './components/Drawer/cod-drawer.js';
 import './components/Button/cod-button.js';
+import './components/ButtonGroup/cod-button-group.js';
